# api-definitions/ai-openapi.json
{
  "components": {
    "schemas": {
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseRelative": {
        "properties": {
          "level": {
            "type": "integer",
            "format": "int32"
          },
          "siblings": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "parents": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        },
        "required": [
          "level",
          "siblings",
          "parents"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseCrossRef": {
        "properties": {
          "crossReference": {
            "type": "string"
          },
          "source": {
            "type": "string"
          }
        },
        "required": [
          "crossReference",
          "source"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.partialaddress.AddressByPartialAddressResponseContainer": {
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "dataVersion": {
            "type": "string"
          },
          "response": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.partialaddress.AddressByPartialAddressResponse"
          },
          "status": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError"
            }
          }
        },
        "required": [
          "apiVersion",
          "dataVersion",
          "response",
          "status",
          "errors"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.partialaddress.AddressByPartialAddressResponse": {
        "properties": {
          "input": {
            "type": "string"
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressNonIDS"
            }
          },
          "filter": {
            "type": "string"
          },
          "fallback": {
            "type": "boolean"
          },
          "historical": {
            "type": "boolean"
          },
          "epoch": {
            "type": "string"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "maxScore": {
            "type": "number",
            "format": "double"
          },
          "verbose": {
            "type": "boolean"
          },
          "highlight": {
            "type": "string"
          },
          "favourpaf": {
            "type": "boolean"
          },
          "favourwelsh": {
            "type": "boolean"
          },
          "countryBoosts": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.CountryBoosts"
          },
          "timeout": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "input",
          "addresses",
          "filter",
          "fallback",
          "historical",
          "epoch",
          "limit",
          "offset",
          "total",
          "maxScore",
          "verbose",
          "highlight",
          "favourpaf",
          "favourwelsh",
          "countryBoosts",
          "timeout"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePaf": {
        "properties": {
          "udprn": {
            "type": "string"
          },
          "organisationName": {
            "type": "string"
          },
          "departmentName": {
            "type": "string"
          },
          "subBuildingName": {
            "type": "string"
          },
          "buildingName": {
            "type": "string"
          },
          "buildingNumber": {
            "type": "string"
          },
          "dependentThoroughfare": {
            "type": "string"
          },
          "thoroughfare": {
            "type": "string"
          },
          "doubleDependentLocality": {
            "type": "string"
          },
          "dependentLocality": {
            "type": "string"
          },
          "postTown": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "postcodeType": {
            "type": "string"
          },
          "deliveryPointSuffix": {
            "type": "string"
          },
          "welshDependentThoroughfare": {
            "type": "string"
          },
          "welshThoroughfare": {
            "type": "string"
          },
          "welshDoubleDependentLocality": {
            "type": "string"
          },
          "welshDependentLocality": {
            "type": "string"
          },
          "welshPostTown": {
            "type": "string"
          },
          "poBoxNumber": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "udprn",
          "organisationName",
          "departmentName",
          "subBuildingName",
          "buildingName",
          "buildingNumber",
          "dependentThoroughfare",
          "thoroughfare",
          "doubleDependentLocality",
          "dependentLocality",
          "postTown",
          "postcode",
          "postcodeType",
          "deliveryPointSuffix",
          "welshDependentThoroughfare",
          "welshThoroughfare",
          "welshDoubleDependentLocality",
          "welshDependentLocality",
          "welshPostTown",
          "poBoxNumber",
          "startDate",
          "endDate"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseNag": {
        "properties": {
          "uprn": {
            "type": "string"
          },
          "postcodeLocator": {
            "type": "string"
          },
          "addressBasePostal": {
            "type": "string"
          },
          "usrn": {
            "type": "string"
          },
          "lpiKey": {
            "type": "string"
          },
          "pao": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePao"
          },
          "sao": {
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseSao"
          },
          "level": {
            "type": "string"
          },
          "officialFlag": {
            "type": "string"
          },
          "logicalStatus": {
            "type": "string"
          },
          "streetDescriptor": {
            "type": "string"
          },
          "townName": {
            "type": "string"
          },
          "locality": {
            "type": "string"
          },
          "organisation": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "localCustodianCode": {
            "type": "string"
          },
          "localCustodianName": {
            "type": "string"
          },
          "localCustodianGeogCode": {
            "type": "string"
          },
          "lpiEndDate": {
            "type": "string"
          },
          "lpiStartDate": {
            "type": "string"
          }
        },
        "required": [
          "uprn",
          "postcodeLocator",
          "addressBasePostal",
          "usrn",
          "lpiKey",
          "pao",
          "sao",
          "level",
          "officialFlag",
          "logicalStatus",
          "streetDescriptor",
          "townName",
          "locality",
          "organisation",
          "legalName",
          "localCustodianCode",
          "localCustodianName",
          "localCustodianGeogCode",
          "lpiEndDate",
          "lpiStartDate"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseAddressNonIDS": {
        "properties": {
          "uprn": {
            "type": "string"
          },
          "parentUprn": {
            "type": "string"
          },
          "relatives": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseRelative"
            }
          },
          "crossRefs": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseCrossRef"
            }
          },
          "formattedAddress": {
            "type": "string"
          },
          "formattedAddressNag": {
            "type": "string"
          },
          "formattedAddressPaf": {
            "type": "string"
          },
          "welshFormattedAddressNag": {
            "type": "string"
          },
          "welshFormattedAddressPaf": {
            "type": "string"
          },
          "highlights": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlight"
          },
          "paf": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponsePaf"
          },
          "nag": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseNag"
            }
          },
          "geo": {
            "nullable": true,
            "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseGeo"
          },
          "classificationCode": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "lpiLogicalStatus": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number",
            "format": "double"
          },
          "underlyingScore": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "uprn",
          "parentUprn",
          "formattedAddress",
          "formattedAddressNag",
          "formattedAddressPaf",
          "welshFormattedAddressNag",
          "welshFormattedAddressPaf",
          "classificationCode",
          "countryCode",
          "lpiLogicalStatus",
          "confidenceScore",
          "underlyingScore"
        ]
      },
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlight": {
        "properties": {
          "bestMatchAddress": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "lang": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.address.AddressResponseHighlightHit"
            }
          }
        },
        "required": [
          "bestMatchAddress",
          "source",
          "lang"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseError": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.AddressResponseStatus": {
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "uk.gov.ons.addressIndex.model.server.response.address.CountryBoosts": {
        "properties": {
          "eboost": {
            "type": "number",
            "format": "double"
          },
          "nboost": {
            "type": "number",
            "format": "double"
          },
          "sboost": {
            "type": "number",
            "format": "double"
          },
          "wboost": {
            "type": "number",
            "format": "double"
          },
          "lboost": {
            "type": "number",
            "format": "double"
          },
          "mboost": {
            "type": "number",
            "format": "double"
          },
          "jboost": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "eboost",
          "nboost",
          "sboost",
          "wboost",
          "lboost",
          "mboost",
          "jboost"
        ]
      },
...
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/addresses/partial": {
      "get": {
        "operationId": "partialAddressQuery",
        "tags": [
          "routes"
        ],
        "summary": "Search by partial address (for type ahead).",
        "description": "Returns a list addresses, from a partial match on the input string",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "input",
            "description": "Specifies the address input.",
            "required": true
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "fallback",
            "description": "Specifies whether a slow fallback query is used in the event of the main query returning no results.",
            "required": false
          },
          {
            "schema": {
              "default": 0,
              "nullable": true,
              "maximum": 250,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "description": "Specifies the offset from zero, used for pagination.",
            "required": false
          },
          {
            "schema": {
              "default": 10,
              "nullable": true,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "description": "Specifies the number of addresses to return.",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "classificationfilter",
            "description": "Classification code filter. Can be pattern match (ZW*), exact match (RD06), multiple exact match (RD02,RD04) or a preset keyword such as residential, commercial, educational or workplace",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "historical",
            "description": "Include historical addresses",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "verbose",
            "description": "Include the full address details in the response (including relatives, crossRefs, paf and nag).",
            "required": false
          },
          {
            "schema": {
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "epoch",
            "description": "Select a specific AddressBase Epoch to search.",
            "required": false
          },
          {
            "schema": {
              "default": "on",
              "nullable": true,
              "type": "string"
            },
            "in": "query",
            "name": "highlight",
            "description": "include detailed highlighting info, options are off, on or debug",
            "required": false
          },
          {
            "schema": {
              "default": true,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "favourpaf",
            "description": "paf beats nag on draw for best match",
            "required": false
          },
          {
            "schema": {
              "default": false,
              "nullable": true,
              "type": "boolean"
            },
            "in": "query",
            "name": "favourwelsh",
            "description": "Welsh beats English on draw for best match",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "eboost",
            "description": "Weighting for addresses in England as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "nboost",
            "description": "Weighting for addresses in Northern Ireland as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "sboost",
            "description": "Weighting for addresses in Scotland as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "wboost",
            "description": "Weighting for addresses in Wales as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "lboost",
            "description": "Weighting for addresses in Channel Islands as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "mboost",
            "description": "Weighting for addresses in Isle of Man as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "nullable": true,
              "format": "decimal",
              "type": "number"
            },
            "in": "query",
            "name": "jboost",
            "description": "Weighting for addresses outside country boundaries as a decimal from 0 to 10",
            "required": false
          },
          {
            "schema": {
              "default": 250,
              "nullable": true,
              "maximum": 5000,
              "type": "integer"
            },
            "in": "query",
            "name": "timeout",
            "description": "Number of milliseconds to allow query to run before bailing out with results so far",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Success. A json return of matched addresses.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/uk.gov.ons.addressIndex.model.server.response.partialaddress.AddressByPartialAddressResponseContainer"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Indicates an issue with the request. Further details are provided in the response."
          },
          "401": {
            "description": "Unauthorised. The API key provided with the request is invalid."
          },
          "429": {
            "description": "Server too busy. The Address Index API is experiencing exceptional load."
          },
          "500": {
            "description": "Internal server error. Failed to process the request due to an internal error."
          }
        }
      }
    },
...
  },
  "info": {
    "title": "ONS Address Index API",
    "description": "The ONS Address Index API lets you retrieve the official version of addresses in England and Wales. The data is currently updated on a 6 week rolling basis.",
    "version": "see /version below"
  },
...
}